Move trait and straighten out result types
